Add properties for Safari browsers
Safari does not support manifest.json file. It will not fail to open the app, but
it will simply ignore it.
That's why we need to add some properties to the .html files, in order to get a comparable
feeling of chrome browser exprience.
To make sure the app looks also good on iOS and Safari, some meta tags are added.